+++ js/jquery/Toolkit.js
@@ -511,17 +511,25 @@ if (!$.event.special.clickout) {
             }
 
             var trigger = true,
-                collection = $(document);
+                collection = $(document),
+                target = $(e.target);
 
             $.each(elements, function(i, item) {
-                if (trigger) {
-                    var self = $(item);
+                var self = $(item);
 
-                    if (!self.is(e.target) && !self.has(e.target).length) {
-                        collection = collection.add(item);
-                    } else {
-                        trigger = false;
-                    }
+                // Test that the delegated selector class matches
+                if ($.type(item) === 'string') {
+                    trigger = (!target.is(item) && !self.has(e.target).length);
+
+                // Else test if the element matches
+                } else {
+                    trigger = (!self.is(e.target) && !self.has(e.target).length);
+                }
+
+                if (trigger) {
+                    collection = collection.add(self);
+                } else {
+                    return false;
                 }
             });